Gradle 2.14: dependencies.gradleApi() gir ikke transitive avhengigheter lengre i runtime. Legger disse til eksplisitt.
"Implementing a Gradle plugin requires use of the gradleApi() dependency in order to compile against Gradle classes. Previously, this encompassed the entire Gradle runtime including Gradle's third party dependencies (e.g. Guava). If the plugin depended on a library that Gradle also depended on, the behavior was unpredictable as multiple versions of classes could end up on the classpath. This would often manifest as unexpected compile errors or runtime linking errors.

Starting with Gradle 2.14, gradleApi() no longer exposes Gradle's implementation dependencies. As such, the compile and test classes for plugins have changed with this release."

[git-p4: depot-paths = "//sktools/SKToolsKode/trunk/": change = 71865]